requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
opencv-python>=4.8
matplotlib>=3.7
Pillow>=10.0
